# Conflicts:
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/schedule/dayin/data/mainD/MoneyDb.kt
@@ -10,7 +10,7 @@ data class MoneyDb(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0L, //식별자
     var date: Date, //등록 날짜
-    var money: Int, //금액
+    var money: Long, //금액
     var auto: Int = 0, //자동 등록 여부 0:기본(등록안함) 1:매주 2:매달 3:매년
     var memo: String? = null, //메모
     var pic: Bitmap? = null, //사진
